requirements.txt
allure_python_commons==2.9.45
Faker==37.5.3
playwright==1.54.0
pytest==7.0.0
